Add files via upload
Yes, you read this correctly. I was inspired by a post on reddit yesterday. So, I set out to create a custom, dialed in workflow that turns Wan 2.1 into an image generator. The simple thing to do is make the frames 1 but, that will not make you happy and give you a reason to continue using it like that. In comes, my workflow. This workflow is optimized for 16gb gpu but I would like to see if any of you are able to load all the models below 16gb. 

This is a magic trick right before your eyes. A video Model that acts like an Image Model that uses SDXL for the detailer. I know this workflow is not going to change the world but, when you love to tweak and create custom ideas, this is way too much fun. To make things interesting, this is not just Wan 2.1 base. I used Wan2.1 Fusionx as the model. 

The settings that I have on this workflow probably should be left as is. I, mean you can mess with anything you want but just be careful.